// frontend/src/lib/api/models/PracticeAnswerFeedback.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PracticeAnswerFeedback = {
    entryId: string;
    isCorrect: (boolean | null);
    correctEntryId: (string | null);
    term: string;
    definition: (string | null);
    outcome: (string | null);
};