// frontend/src/lib/api/models/PracticeResult.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PracticeResult = {
    sessionId: string;
    status: string;
    totalCount: (number | null);
    correctCount: (number | null);
    notKnownCount: (number | null);
};